For some reason the test wont run in CI, moving to a test file that for sure runs in CI?
