+++ src/EditorFeatures/Test/CommentSelection/ToggleBlockCommentCommandHandlerTests.cs
@@ -5,9 +5,14 @@
 #nullable disable
 
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Editor.Implementation.CommentSelection;
 using Microsoft.CodeAnalysis.Editor.UnitTests.Workspaces;
+using Microsoft.CodeAnalysis.ErrorReporting;
+using Microsoft.CodeAnalysis.Remote;
+using Microsoft.CodeAnalysis.Remote.Testing;
 using Microsoft.CodeAnalysis.Test.Utilities;
 using Microsoft.CodeAnalysis.Test.Utilities.CommentSelection;
 using Microsoft.VisualStudio.Commanding;
@@ -20,6 +25,32 @@ namespace Microsoft.CodeAnalysis.Editor.UnitTests.CommentSelection
     [UseExportProvider]
     public class ToggleBlockCommentCommandHandlerTests : AbstractToggleCommentTestBase
     {
+        [Fact, Trait("NonFatalWatsonThrowsTrait", "NonFatalWatsonThrowsValue")]
+        public async Task Throws()
+        {
+            Console.WriteLine("Running test");
+            var localComposition = EditorTestCompositions.EditorFeatures.WithTestHostParts(Remote.Testing.TestHost.OutOfProcess);
+            using var localWorkspace = new TestWorkspace(composition: localComposition);
+
+            var clientProvider = (InProcRemoteHostClientProvider?)localWorkspace.Services.GetService<IRemoteHostClientProvider>();
+            Assert.NotNull(clientProvider);
+
+            var client = await InProcRemoteHostClient.GetTestClientAsync(localWorkspace).ConfigureAwait(false);
+            var remoteWorkspace = client.TestData.WorkspaceManager.GetWorkspace();
+            ServiceBase.Equals(1, 1);
+            var tasks = new List<Task>();
+            for (var i = 0; i < 10; i++)
+            {
+                tasks.Add(Task.Run(() =>
+                {
+                    Console.WriteLine("fail " + i);
+                    FatalError.ReportAndCatch(new Exception("TestException"));
+                }));
+            }
+
+            await Task.WhenAll(tasks).ConfigureAwait(false);
+        }
+
         [WpfFact, Trait(Traits.Feature, Traits.Features.ToggleBlockComment)]
         public void AddComment_EmptyCaret()
         {
